@@ -2,25 +2,13 @@ package net.minevn.flexitem;
 
 import net.minevn.flexitem.object.FlexContainer;
 
-import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 
 public class Storage {
 
-    private static Storage instance;
-
-    private final ConcurrentHashMap<UUID, FlexContainer> containers;
-
-    public Storage() {
-        instance = this;
-        containers = new ConcurrentHashMap<>();
-    }
-
-    public static Storage getInstance() {
-        return instance;
-    }
+    private final ConcurrentHashMap<UUID, FlexContainer> containers = new ConcurrentHashMap<>();
 
     public Map<UUID, FlexContainer> getContainers() {
         return containers;
@@ -37,6 +25,4 @@ public class Storage {
     public void addContainer(FlexContainer container) {
         containers.put(container.getUUID(), container);
     }
-
-
 }
