@@ -13,10 +13,18 @@ import org.bukkit.entity.Player;
 
 import java.util.UUID;
 
+import static net.minevn.flexitem.object.TypeContainer.Armor;
+import static net.minevn.flexitem.object.TypeContainer.Hand;
+
 public class Utils {
 
-    public static void sendMessages(Player player, UUID uuid) {
-        TextComponent message = new TextComponent(player.getDisplayName() + " vừa khoe 1 cái gì đó");
+    public static void sendMessages(Player player, UUID uuid, TypeContainer type) {
+        String typeString = switch (type) {
+            case Armor -> "trang bị";
+            case Hand -> "vật phẩm";
+            default -> "bộ vật phẩm";
+        };
+        TextComponent message = new TextComponent(player.getDisplayName() + " vừa khoe " + typeString + " của mình, click để xem");
         message.setClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/flexitem open " + uuid));
         for (Player online : Bukkit.getOnlinePlayers()) {
             online.spigot().sendMessage(message);
@@ -25,26 +33,25 @@ public class Utils {
 
     public static void openContainer(FlexContainer container) {
         var player = Bukkit.getPlayer(container.getOwner());
-        if (container.getType() == TypeContainer.Armor) {
-            ChestGui gui = new ChestGui(4, "Khoe vật phẩm của " + player.getDisplayName());
-            StaticPane pane = new StaticPane(0, 0, 9, 6);
-            int x = 4, y = 0;
+        if (container.getType() == Armor) {
+            ChestGui gui = new ChestGui(3, "Khoe vật phẩm của " + player.getDisplayName());
+            StaticPane pane = new StaticPane(0, 0, 9, 3);
+            int x = 1;
             for (var item : container.getItems()) {
-                Bukkit.broadcastMessage(item.getType().name() + " x: " + x + " y: " + y + "");
-                pane.addItem(new GuiItem(item), x, y);
-                x += 9;
-                y++;
+                Bukkit.broadcastMessage(item.getType().name() + " x: " + x);
+                pane.addItem(new GuiItem(item), x, 1);
+                x += 2;
             }
             pane.setOnClick(event -> event.setCancelled(true));
             gui.addPane(pane);
             gui.show(player);
             return;
         }
-        if (container.getType() == TypeContainer.Hand) {
+        if (container.getType() == Hand) {
             DropperGui gui = new DropperGui("Khoe vật phẩm của " + player.getDisplayName());
             gui.setOnGlobalClick(event -> event.setCancelled(true));
-            StaticPane pane = new StaticPane(0, 0, 3, 9);
-            pane.addItem(new GuiItem(container.getItems().get(0)), 5, 1);
+            StaticPane pane = new StaticPane(1, 1, 1, 1);
+            pane.addItem(new GuiItem(container.getItems().get(0)), 0, 0);
             gui.getContentsComponent().addPane(pane);
             gui.show(player);
             return;
